+++ cpp/src/db/engine/ExecutionEngineImpl.cpp
@@ -1,22 +1,33 @@
-/*******************************************************************************
- * Copyright 上海赜睿信息科技有限公司(Zilliz) - All Rights Reserved
- * Unauthorized copying of this file, via any medium is strictly prohibited.
- * Proprietary and confidential.
- ******************************************************************************/
-#include <stdexcept>
-#include "src/cache/GpuCacheMgr.h"
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 
-#include "src/metrics/Metrics.h"
-#include "db/Log.h"
-#include "utils/CommonUtil.h"
-
-#include "src/cache/CpuCacheMgr.h"
 #include "ExecutionEngineImpl.h"
+#include "cache/GpuCacheMgr.h"
+#include "cache/CpuCacheMgr.h"
+#include "metrics/Metrics.h"
+#include "utils/Log.h"
+#include "utils/CommonUtil.h"
+#include "utils/Exception.h"
+
 #include "wrapper/knowhere/vec_index.h"
 #include "wrapper/knowhere/vec_impl.h"
 #include "knowhere/common/exception.h"
-#include "db/Exception.h"
 
+#include <stdexcept>
 
 namespace zilliz {
 namespace milvus {
@@ -124,7 +135,7 @@ Status ExecutionEngineImpl::Serialize() {
 }
 
 Status ExecutionEngineImpl::Load(bool to_cache) {
-    index_ = zilliz::milvus::cache::CpuCacheMgr::GetInstance()->GetIndex(location_);
+    index_ = cache::CpuCacheMgr::GetInstance()->GetIndex(location_);
     bool already_in_cache = (index_ != nullptr);
     if (!already_in_cache) {
         try {
@@ -151,7 +162,7 @@ Status ExecutionEngineImpl::Load(bool to_cache) {
 }
 
 Status ExecutionEngineImpl::CopyToGpu(uint64_t device_id) {
-    auto index = zilliz::milvus::cache::GpuCacheMgr::GetInstance(device_id)->GetIndex(location_);
+    auto index = cache::GpuCacheMgr::GetInstance(device_id)->GetIndex(location_);
     bool already_in_cache = (index != nullptr);
     if (already_in_cache) {
         index_ = index;
@@ -178,7 +189,7 @@ Status ExecutionEngineImpl::CopyToGpu(uint64_t device_id) {
 }
 
 Status ExecutionEngineImpl::CopyToCpu() {
-    auto index = zilliz::milvus::cache::CpuCacheMgr::GetInstance()->GetIndex(location_);
+    auto index = cache::CpuCacheMgr::GetInstance()->GetIndex(location_);
     bool already_in_cache = (index != nullptr);
     if (already_in_cache) {
         index_ = index;
@@ -221,7 +232,7 @@ Status ExecutionEngineImpl::Merge(const std::string &location) {
     }
     ENGINE_LOG_DEBUG << "Merge index file: " << location << " to: " << location_;
 
-    auto to_merge = zilliz::milvus::cache::CpuCacheMgr::GetInstance()->GetIndex(location);
+    auto to_merge = cache::CpuCacheMgr::GetInstance()->GetIndex(location);
     if (!to_merge) {
         try {
             double physical_size = server::CommonUtil::GetFileSize(location);