+++ cpp/src/scheduler/task/SearchTask.cpp
@@ -1,14 +1,25 @@
-/*******************************************************************************
- * Copyright 上海赜睿信息科技有限公司(Zilliz) - All Rights Reserved
- * Unauthorized copying of this file, via any medium is strictly prohibited.
- * Proprietary and confidential.
- ******************************************************************************/
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 
-#include "src/metrics/Metrics.h"
-#include "src/utils/TimeRecorder.h"
-#include "src/db/engine/EngineFactory.h"
-#include "src/db/Log.h"
 #include "SearchTask.h"
+#include "metrics/Metrics.h"
+#include "db/engine/EngineFactory.h"
+#include "utils/TimeRecorder.h"
+#include "utils/Log.h"
 
 #include <thread>
 
@@ -97,7 +108,7 @@ XSearchTask::XSearchTask(TableFileSchemaPtr file)
 
 void
 XSearchTask::Load(LoadType type, uint8_t device_id) {
-    server::TimeRecorder rc("");
+    TimeRecorder rc("");
     Status stat = Status::OK();
     std::string error_msg;
     std::string type_str;
@@ -166,7 +177,7 @@ XSearchTask::Execute() {
     ENGINE_LOG_DEBUG << "Searching in file id:" << index_id_ << " with "
                      << search_contexts_.size() << " tasks";
 
-    server::TimeRecorder rc("DoSearch file id:" + std::to_string(index_id_));
+    TimeRecorder rc("DoSearch file id:" + std::to_string(index_id_));
 
     server::CollectDurationMetrics metrics(index_type_);