+++ cpp/CHANGELOG.md
@@ -8,10 +8,14 @@ Please mark all change in change log and use the ticket from JIRA.
 
 ## Improvement
 - MS-552 - Add and change the easylogging library
+- MS-553 - Refine cache code
+- MS-557 - Merge Log.h
+- MS-556 - Add Job Definition in Scheduler
 
 ## New Feature
 
 ## Task
+- MS-554 - Change license to Apache 2.0
 
 # Milvus 0.4.0 (2019-09-12)